factor-trading/src/dags/backtest_framework_example/requirements.txt
apache-airflow-providers-amazon>=2.0.0
boto3>=1.24.0
pandas>=1.3.0
matplotlib>=3.5.0
seaborn>=0.11.0
pyyaml>=6.0